refactor: centralize pdf generation
